refactor(channels): extract telegram parse + transport modules (#162)
Move the Telegram parsing and HTTP transport out of the monolithic
channels/telegram.ts into:

- src/channels/telegram/parse.ts  — TelegramMessage / TelegramAttachment /
  TelegramReplyTo / raw update types, extractAttachment, extractReplyTo,
  parseGetUpdatesResult, formatAttachmentUserText, inboxDir,
  extensionFromMime, synthesizeFileName, sanitizeEnvelopeField,
  formatReplyToContext, REPLY_TO_SNIPPET_MAX, normalizeChatType,
  stripBotMention, TELEGRAM_BOT_DOWNLOAD_CAP_BYTES.
- src/channels/telegram/transport.ts — TelegramTransport (extends
  ChannelTransport), HttpTelegramTransport, control/download timeout
  constants, guessMimeFromPath.

channels/telegram.ts becomes a barrel: it re-exports every moved name under
its original identifier so all importers and tests compile unchanged. The
engine/routing/prompts remain inline for now (moved in later phases).

Verbatim move — no logic changes. Full suite: 1184 pass / 0 fail (baseline).
